split the role map file, improve reading in of the configmap
